backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
mysql-connector-python
oracledb
pandas
openpyxl
requests
redis
python-jose[cryptography]
python-multipart
email-validator
passlib[bcrypt]
